pointers are output as lists
a pointer message is really a "list <ptr>" rather than
"pointer <ptr>" (which i would have guessed due to the
existance of 's_pointer').
this can be confirmed by sending a pointer via [t a] to the
right inlet of [set] (which gives the error:
  "inlet: expected 'pointer' but got 'pointer'"

checking back in m_class.c, it seems that the pointer-method
will only ever get called from within the pd_defaultlist.